feat: virtualise content panel, add initial dynamic content modal
