+++ src/components/objectIdentifier/index.tsx
@@ -1,0 +1,4 @@
+export * from "./card/objectIdentifierCard.component";
+export * from "./card/contentObjectIdentifierCard.component";
+export * from "./list/objectIdentifierList.component";
+export * from "./list/sortableContentObjectIdentifierList.component";
