--- src/components/objectIdentifierCard/index.tsx
@@ -1,1 +1,0 @@
-export * from "./objectIdentifier.component";
