+++ src/components/panel/panelSections/panelPlayback.component.tsx
@@ -12,7 +12,7 @@ import {
   IntegrationUploaderProvider,
 } from "src/components/integrations";
 import { DisplayGraphQLQuery } from "src/components/modals";
-import { ObjectIdentifierCard } from "src/components/objectIdentifierCard";
+import { ObjectIdentifierCard } from "src/components/objectIdentifier";
 import { pollPanelRefetch } from "src/components/panel/panel.lib";
 import { PanelLoading } from "src/components/panel/panelLoading";
 import {
